Enable usage of keyboard shortcts without mixins
